# documents/swagger20.yaml
...
swagger: "2.0"
info:
  description: "No description"
  version: "1.0.0"
  title: "NaplesPizza"
consumes:
- "application/json"
produces:
- "application/json"
paths:
...
  /orders:
    get:
      summary: "Get the order by customerid/order date"
      parameters:
      - name: "customerId"
        in: "query"
        required: false
        type: "string"
        description: "Order By CustomerID"
      - name: "storeId"
        in: "query"
        required: false
        type: "string"
        description: "Order By Store Id"
      - name: "orderDate"
        in: "query"
        required: false
        type: "string"
      responses:
        200:
          description: "Status 200"
    post:
      summary: "Create new order"
      description: "Tp created new order composing of storeid and customerid"
      parameters: []
      responses:
        200:
          description: "Status 200"
  /orders/{orderId}:
    put:
      summary: "Update the order status, shipping and payment information"
      parameters: []
      responses:
        200:
          description: "Status 200"
    parameters:
    - name: "orderId"
      in: "path"
      required: true
      type: "string"
...